Moving around settings Navigation tests
